add millisecond wait for condition

+++ muduo/base/AsyncLogging.cc
@@ -71,7 +71,7 @@ void AsyncLogging::threadFunc()
       muduo::MutexLockGuard lock(mutex_);
       if (buffers_.empty())  // unusual usage!
       {
-        cond_.waitForSeconds(flushInterval_);
+        cond_.waitForSeconds(static_cast<double>(flushInterval_));
       }
       buffers_.push_back(currentBuffer_.release());
       currentBuffer_ = boost::ptr_container::move(newBuffer1);
